requirements.txt
attrs==20.2.0
bumpversion==0.5.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
checksumdir==1.2.0
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3; sys_platform == "win32"
colorlog==4.2.1
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
dynaconf==3.0.0
fiona==1.8.17
geoalchemy2==0.6.3
geojson==2.5.0
geopandas==0.7.0
gitchangelog==3.0.4
idna==2.10
invoke==1.4.1
kiwisolver==1.2.0
matplotlib==3.3.2
more-itertools==8.5.0
munch==2.5.0
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
ortools==7.8.7959
pandas==1.1.2
pillow==7.2.0
protobuf==3.13.0
pycparser==2.20
pyparsing==2.4.7
pyproj==2.6.1.post1
python-box==5.1.1
python-dateutil==2.8.1
python-log-indenter==0.9
python-telegram-bot==12.8
pytz==2020.1
pyyaml==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2; platform_python_implementation == "CPython" and python_version < "3.9"
scipy==1.5.2
shapely==1.7.1
six==1.15.0
sortedcontainers==2.2.2
sqlalchemy==1.3.19
toml==0.10.1
tornado==6.0.4
urllib3==1.25.10
visilibity==1.0.10
wrapt==1.12.1
